refactor: 레이어 역전 일부 해소 — 공용 유틸/컴포넌트 하위 레이어로 이동 (#138)
app.component·app.module 가 feature 를 import 하던 경계 위반을 순수 파일
이동으로 해소한다. 동작 변경 없음.

- diaryImageUrl 유틸을 feature/diary/shared/utils → app.module/utils 로 이동
  (@module/utils/url 에만 의존하는 순수 유틸). importer 17곳 경로 갱신.
  AppBottomNav·NativeBridge·serverApi 및 member→diary 결합 일부 감소.
- ChallengeChip 을 feature/challenge/shared/components → app.component/cards
  로 이동(순수 프레젠테이션, feature 의존 없음). 유일 importer DiaryCard 갱신.
- useTokenRefreshOnResume 이 authApi.refreshToken() 대신 동일 구현인
  @module/api/tokenRefresh 의 refreshAccessTokenOnce() 를 직접 사용.

남은 역전(별도 처리 필요, 이번 범위 제외):
- 앱 셸(AppBottomNav·useAuthLayoutState·AppLayoutShell)의 member/notification
  훅·프롬프트 의존: 훅 인스턴스화 위치/데이터 흐름을 바꿔야 해(props 주입
  또는 shell 레이어 분리) 순수 이동이 아니며 회귀 위험이 있어 보류.
- serverApi 의 DiaryItem 타입 의존: 타입 전용 역전으로 영향이 작아 보류.
