+++ src/app.component/cards/ChallengeChip.tsx
